+++ src/Lesson_1/Marafon/Course/Cross.java
@@ -1,16 +1,16 @@
-package Lesson_1.Marafon;
+package Lesson_1.Marafon.Course;
 
 import Lesson_1.Marafon.Competitors.Competitor;
 
-public class Water extends Obstacle {
+public class Cross extends Obstacle {
     int length;
 
-    public Water(int length) {
+    public Cross(int length) {
         this.length = length;
     }
 
     @Override
     public void doIt(Competitor competitor) {
-        competitor.swim(length);
+        competitor.run(length);
     }
 }
