+++ src/Lesson_1/Marafon/Course/Obstacle.java
@@ -1,4 +1,4 @@
-package Lesson_1.Marafon;
+package Lesson_1.Marafon.Course;
 
 import Lesson_1.Marafon.Competitors.Competitor;
