+++ src/Lesson_1/Marafon/Main.java
@@ -4,6 +4,9 @@ import Lesson_1.Marafon.Competitors.Cat;
 import Lesson_1.Marafon.Competitors.Competitor;
 import Lesson_1.Marafon.Competitors.Dog;
 import Lesson_1.Marafon.Competitors.Human;
+import Lesson_1.Marafon.Course.Cross;
+import Lesson_1.Marafon.Course.Obstacle;
+import Lesson_1.Marafon.Course.Wall;
 
 public class Main {
     public static void main(String[] args) {
